persist user secrets in indexed db
