+++ .circleci/config.yml
@@ -55,7 +55,7 @@ jobs:
           key: offen-vault-{{ checksum "package.json" }}
       - run:
           name: Run tests
-          command: npm test -- --allow-chrome-as-root
+          command: npm test
 
 workflows:
   version: 2
